Revert "Add Labels API for `labels-str` and `labels-file` (#52085)"
This reverts commit 4d28121c74bbd9e0fd9c18bfa3958ee5f7975e68.
