+++ python/ray/scripts/scripts.py
@@ -25,15 +25,11 @@ import yaml
 import ray
 import ray._private.ray_constants as ray_constants
 import ray._private.services as services
-from ray._private.label_utils import (
-    parse_node_labels_json,
-    parse_node_labels_from_yaml_file,
-    parse_node_labels_string,
-)
 from ray._private.utils import (
     check_ray_client_dependencies_installed,
     load_class,
     parse_resources_json,
+    parse_node_labels_json,
 )
 from ray._private.internal_api import memory_summary
 from ray._private.usage import usage_lib
@@ -638,19 +634,9 @@ Windows powershell users need additional escaping:
     "--labels",
     required=False,
     hidden=True,
-    default="",
+    default="{}",
     type=str,
-    help="a string list of key-value pairs mapping label name to label value."
-    "These values take precedence over conflicting keys passed in from --labels-file."
-    'Ex: --labels "key1=val1,key2=val2"',
-)
-@click.option(
-    "--labels-file",
-    required=False,
-    hidden=True,
-    default="",
-    type=str,
-    help="a path to a YAML file containing a dictionary mapping of label keys to values.",
+    help="a JSON serialized dictionary mapping label name to label value.",
 )
 @click.option(
     "--include-log-monitor",
@@ -748,7 +734,6 @@ def start(
     ray_debugger_external,
     disable_usage_stats,
     labels,
-    labels_file,
     include_log_monitor,
     enable_resource_isolation,
     system_reserved_cpu,
@@ -771,42 +756,7 @@ def start(
         node_ip_address = services.resolve_ip_for_localhost(node_ip_address)
 
     resources = parse_resources_json(resources, cli_logger, cf)
-
-    # Compose labels passed in with `--labels` and `--labels-file`.
-    # In the case of duplicate keys, the values from `--labels` take precedence.
-    try:
-        labels_from_file = parse_node_labels_from_yaml_file(labels_file)
-    except Exception as e:
-        cli_logger.abort(
-            "The file at `{}` is not a valid YAML file, detailed error: {} "
-            "Valid values look like this: `{}`",
-            cf.bold(f"--labels-file={labels_file}"),
-            str(e),
-            cf.bold("--labels-file='gpu_type: A100\nregion: us'"),
-        )
-    try:
-        # Attempt to parse labels from new string format first.
-        labels_from_string = parse_node_labels_string(labels)
-    except Exception as e:
-        try:
-            # Fall back to JSON format if parsing from string fails.
-            labels_from_string = parse_node_labels_json(labels)
-            warnings.warn(
-                "passing node labels with `--labels` in JSON format is "
-                "deprecated and will be removed in a future version of Ray.",
-                DeprecationWarning,
-                stacklevel=2,
-            )
-        except Exception:
-            # If parsing labels from both formats fails, return the original error message.
-            cli_logger.abort(
-                "`{}` is not a valid string of key-value pairs, detailed error: {} "
-                "Valid values look like this: `{}`",
-                cf.bold(f"--labels={labels}"),
-                str(e),
-                cf.bold('--labels="key1=val1,key2=val2"'),
-            )
-    labels_dict = {**labels_from_file, **labels_from_string}
+    labels_dict = parse_node_labels_json(labels, cli_logger, cf)
 
     if plasma_store_socket_name is not None:
         warnings.warn(
