+++ python/ray/tests/BUILD
@@ -346,7 +346,6 @@ py_test_module_list(
 py_test_module_list(
     size = "medium",
     files = [
-        "test_label_utils.py",
         "test_minimal_install.py",
         "test_runtime_env_ray_minimal.py",
         "test_utils.py",
